Refactor: datalist_test
